PJMS-229: Reorganize test packages.
