+++ integration-tests/src/test/java/org/opennms/minion/test/FeatureInstallKarafTest.java
@@ -2,6 +2,7 @@ package org.opennms.minion.test;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.opennms.minion.test.core.SmnnepoKarafTest;
 import org.ops4j.pax.exam.junit.PaxExam;
 import org.ops4j.pax.exam.spi.reactors.ExamReactorStrategy;
 import org.ops4j.pax.exam.spi.reactors.PerMethod;
